requirements.txt
Django==2.1.7
psycopg2-binary
gunicorn
django-allauth